+++ CHANGELOG.md
@@ -14,9 +14,17 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Features
 
-- storage: improve encoding/decoding of SFA ([#66](https://github.com/lamports-dev/alpamayo/pull/66))
-
 ### Breaking
+
+## [0.11.0] - 2025-07-08
+
+### Fixes
+
+- rpc: fix getBlock response ([#67](https://github.com/lamports-dev/alpamayo/pull/67))
+
+### Features
+
+- storage: improve encoding/decoding of SFA ([#66](https://github.com/lamports-dev/alpamayo/pull/66))
 
 ## [0.11.0] - 2025-06-10